# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Goat Nutrition App API
  version: 0.1.0
  description: |
    Minimal OpenAPI for key endpoints. Auth uses cookie session (Flask-Login).
servers:
  - url: http://localhost:5001
paths:
...
  /api/agent/tip:
    get:
      summary: Daily tip (requires X-Api-Key)
      parameters:
        - in: header
          name: X-Api-Key
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '401': { description: Missing API key }
  /api/agent/recommendation:
    post:
      summary: Nutrition recommendation
      parameters:
        - in: header
          name: X-Api-Key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                EarNum: { type: string }
                Breed: { type: string }
                Body_Weight_kg: { type: number }
                Age_Months: { type: integer }
                Sex: { type: string }
                status: { type: string }
                target_average_daily_gain_g: { type: number }
                milk_yield_kg_day: { type: number }
                milk_fat_percentage: { type: number }
                number_of_fetuses: { type: integer }
                other_remarks: { type: string }
      responses:
        '200': { description: OK }
        '400': { description: Validation error }
  /api/agent/chat:
    post:
      summary: Chat with AI (text or image multipart)
      parameters:
        - in: header
          name: X-Api-Key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message, session_id]
              properties:
                message: { type: string }
                session_id: { type: string }
                ear_num_context: { type: string }
          multipart/form-data:
            schema:
              type: object
              required: [message, session_id]
              properties:
                message: { type: string }
                session_id: { type: string }
                ear_num_context: { type: string }
                image: { type: string, format: binary }
      responses:
        '200': { description: OK }
        '400': { description: Validation error }
...
  /api/agent/analytics-report:
    post:
      summary: Generate analytics report via AI
      parameters:
        - in: header
          name: X-Api-Key
          required: true
          schema:
            type: string
          description: Gemini API key used for report generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsReportRequest'
      responses:
        '200':
          description: Report generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  report_html: { type: string }
                  report_markdown: { type: string }
        '400': { description: Validation error }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    AnalyticsReportRequest:
      type: object
      properties:
        filters:
          type: object
          additionalProperties: true
        cohort:
          type: array
          items:
            type: object
            additionalProperties: true
        cost_benefit:
          type: object
          additionalProperties: true
        insights:
          type: array
          items: { type: string }
...
security:
  - cookieAuth: []